requirements.txt
torch
torchvision
transformers
huggingface_hub
streamlit
pillow
scikit-learn
tqdm
